requirements.txt
# Core dependencies
fastapi==0.103.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn==0.23.2
python-dotenv==1.0.0
streamlit==1.26.0
pyyaml>=6.0

# Database
sqlalchemy>=1.4.42,<1.5
databases==0.8.0
aiosqlite==0.19.0
psycopg2-binary==2.9.7  # For PostgreSQL in production

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# AI/ML
openai>=1.0.0,<2.0.0
langchain>=0.1.80
langchain-community>=0.0.30
langchain-openai>=0.1.0
langsmith>=0.3.0  # For tracing and debugging

# Utils
requests==2.31.0
uuid==1.30
tenacity==8.2.3  # For API call retries
urllib3==2.0.4
certifi==2023.7.22

# Additional dependencies
faiss-cpu 